+++ docs/HISTORY.rst
@@ -6,6 +6,9 @@ dev
 
 - errors raised in finally clause did not pop up with debug=True
 
+- added utility function ``createFile`` to create file-content the same way
+  as ``createImage`` creates image-content.
+
 0.2.2 (2013-05-08)
 ==================